+++ packages/stream_chat/CHANGELOG.md
@@ -3,6 +3,7 @@
 🐞 Fixed
 
 - [[#710]](https://github.com/GetStream/stream-chat-flutter/issues/710) Fixed JWT requiring using `String` as id.
+- Fixed expired CDN attachment links not updating correctly.
 
 ## 3.0.0